aggregator/requirements.txt
 flask
 requests
 pymongo
 newspaper3k
 lxml
 lxml_html_clean
